refactor: separate reception form into different components
