+++ src/app/(main)/dashboard/abastecimiento/recepciones/components/form/serials-form.tsx
@@ -1,6 +1,5 @@
 'use client'
 import { Input } from '@/modules/common/components/input/input'
-
 import { useFormContext } from 'react-hook-form'
 import {
   FormControl,
@@ -15,24 +14,22 @@ import {
   CardHeader,
   CardTitle,
 } from '@/modules/common/components/card/card'
-import { v4 } from 'uuid'
 import { Switch } from '@/modules/common/components/switch/switch'
 import { nanoid } from 'nanoid'
+
 export function SerialsForm({
   index: indexForm,
   quantity,
   id,
 }: {
+  id: number
   index: number
   quantity: number
-  id?: number
-  closeModal?: () => void
 }) {
   const form = useFormContext()
   const autoSerialsEnabled = form.watch(
     `renglones.${indexForm}.seriales_automaticos`
   )
-
   return (
     <div className="flex flex-col gap-0 p-8 overflow-y-auto">
       <CardHeader className="flex items-center justify-between gap-4">
@@ -49,10 +46,10 @@ export function SerialsForm({
                     checked={field.value}
                     onCheckedChange={(value) => {
                       if (!value) {
-                        Array.from({ length: quantity }).map((_, index) => {
+                        Array.from({ length: quantity }).forEach((_, index) => {
                           form.setValue(
-                            `renglones.${indexForm}.seriales.${index}.serial`,
-                            ''
+                            `renglones.${indexForm}.seriales.${index}`,
+                            {}
                           )
                         })
                       }
@@ -67,40 +64,66 @@ export function SerialsForm({
         />
       </CardHeader>
       {Array.from({ length: quantity }).map((_, index) => (
-        <FormField
+        <div
           key={`renglon-${indexForm}-serial-${index}`}
-          control={form.control}
-          name={`renglones.${indexForm}.seriales.${index}.serial`}
-          rules={{ required: true, validate: (value) => !!value || 'Required' }}
-          render={({ field }) => (
-            <FormItem className="">
-              <FormLabel>Serial #{index + 1}</FormLabel>
-              <FormControl>
-                <Input
-                  {...field}
-                  onChange={(e) => {
-                    if (form.formState.errors[field.name]) {
-                      form.clearErrors(field.name)
+          className="flex flex-row gap-8 mb-4"
+        >
+          <FormField
+            control={form.control}
+            name={`renglones.${indexForm}.seriales.${index}`}
+            rules={{ required: true }}
+            render={({ field }) => (
+              <FormItem className="flex-1 ">
+                <FormLabel>Serial #{index + 1}</FormLabel>
+                <FormControl>
+                  <Input
+                    {...field}
+                    onChange={(e) => {
+                      if (form.formState.errors[field.name]) {
+                        form.clearErrors(field.name)
+                      }
+                      form.setValue(field.name, e.target.value, {
+                        shouldDirty: true,
+                      })
+                    }}
+                    value={
+                      field.value?.serial ||
+                      (autoSerialsEnabled
+                        ? form.setValue(
+                            field.name,
+                            {
+                              serial: nanoid(11),
+                              id_renglon: id,
+                              condicion: null,
+                            },
+                            {
+                              shouldDirty: true,
+                            }
+                          )
+                        : '') ||
+                      ''
                     }
-                    form.setValue(field.name, e.target.value, {
-                      shouldDirty: true,
-                    })
-                  }}
-                  value={
-                    field.value ||
-                    (autoSerialsEnabled
-                      ? form.setValue(field.name, nanoid(11), {
-                          shouldDirty: true,
-                        })
-                      : '') ||
-                    ''
-                  }
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
+                  />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name={`renglones.${indexForm}.seriales.${index}.condicion`}
+            rules={{ required: true }}
+            render={({ field }) => (
+              <FormItem className="flex-1 ">
+                <FormLabel>Condición</FormLabel>
+                <FormControl>
+                  <Input {...field} value={field.value} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
       ))}
     </div>
   )
