+++ src/app/(main)/dashboard/armamento/devoluciones/components/form/serials-form.tsx
@@ -80,7 +80,6 @@ export function SerialsFormNew({
           columns={columns}
           data={serials}
           onSelectedRowsChange={setSelectedData}
-          isColumnFilterEnabled={false}
           selectedData={selectedItems}
           setSelectedData={setSelectedItems}
         />
