+++ src/app/(main)/dashboard/armamento/recepciones/components/form/receptions-form.tsx
@@ -605,7 +605,6 @@ export default function ReceptionsForm({
                     columns={columns}
                     data={renglonesData}
                     onSelectedRowsChange={handleTableSelect}
-                    isColumnFilterEnabled={false}
                     selectedData={selectedRows}
                     isStatusEnabled={false}
                     setSelectedData={setSelectedRows}
